Split app.jsx accross multiple files and upgraded to ES6.
