+++ gulpfile.js
@@ -1,26 +1,29 @@
 var gulp = require('gulp');
-var react = require('gulp-react');
+var browserify = require('browserify');
+var babelify = require('babelify');
+var source = require('vinyl-source-stream');
 
 var paths = {
-  js: ['js/*.js'],
-  jsx: ['js/*.jsx'],
-  copy: ['js/*.js', 'css/*.css', 'js/libs/*.js', 'img/*'],
+  js: ['js/*.jsx'],
+  copy: ['css/*.css', 'js/libs/*.js', 'img/*'],
 };
 
 gulp.task('copy', function() {
   return gulp.src(paths.copy, {base: './'}).pipe(gulp.dest('static'));
 });
 
-gulp.task('react', function() {
-  return gulp.src(paths.jsx)
-    .pipe(react({harmony: true}))
+gulp.task('babel', function() {
+  return browserify({entries: 'js/app.jsx', extensions: ['.jsx'], debug: true})
+    .transform('babelify', {presets: ['es2015', 'react']})
+    .bundle()
+    .pipe(source('bundle.js'))
     .pipe(gulp.dest('static/js'));
 });
 
 gulp.task('watch', function() {
   gulp.watch(paths.copy, ['copy']);
-  gulp.watch(paths.jsx, ['react']);
+  gulp.watch(paths.js, ['babel']);
 });
 
-gulp.task('build', ['react', 'copy']);
+gulp.task('build', ['babel', 'copy']);
 gulp.task('default', ['build', 'watch']);
