+++ index.html
@@ -2,21 +2,24 @@
 <html xmlns:xlink="http://www.w3.org/1999/xlink">
   <head>
     <meta charset="utf-8">
+
     <title>Timberlake :: The 20/20 Experience</title>
+
     <link rel="shortcut icon" href="/static/img/bowtie.png">
+
     <script src="/static/js/libs/jquery-2.1.1.min.js"></script>
     <script src="/static/js/libs/jquery-ui-1.12.1.min.js"></script>
     <script src="/static/js/libs/d3.v3.min.js"></script>
     <script src="/static/js/libs/underscore-min.js"></script>
-    <script src="/static/js/libs/react-0.12.1.min.js"></script>
-    <script src="/static/js/libs/react-router.min.js"></script>
     <script src="/static/js/libs/boxplot.js"></script>
     <script src="/static/js/libs/waterfall.js"></script>
+
     <link rel="stylesheet" href="/static/css/bootstrap.min.css">
     <link rel="stylesheet" href="/static/css/jquery-ui-1.12.1.min.css">
     <link rel="stylesheet" href="/static/css/app.css">
   </head>
   <body>
-    <script src="/static/js/app.js"></script>
+    <div id="timberlake"></div>
+    <script src="/static/js/bundle.js"></script>
   </body>
 </html>
