+++ package.json
@@ -1,19 +1,40 @@
 {
   "name": "timberlake",
-  "version": "1.0.0",
+  "version": "0.0.1",
   "description": "Timberlake is a Job Tracker for Hadoop.",
   "repository": {
     "type": "git",
-    "url": "https://github.com/stripe/timberlake.git"
+    "url": "git+https://github.com/stripe/timberlake.git"
   },
   "author": "Jeff Balogh",
-  "license": "MIT",
   "bugs": {
     "url": "https://github.com/stripe/timberlake/issues"
   },
   "homepage": "https://github.com/stripe/timberlake",
   "devDependencies": {
-    "gulp": "^3.8.10",
-    "gulp-react": "^2.0.0"
+    "babel-plugin-transform-es2015-modules-amd": "^6.18.0",
+    "babel-preset-es2015": "^6.18.0",
+    "babel-preset-react": "^6.16.0",
+    "babelify": "^7.3.0",
+    "browserify": "^13.3.0",
+    "eslint": "^3.13.1",
+    "eslint-config-standard": "^6.2.1",
+    "eslint-plugin-promise": "^3.4.0",
+    "eslint-plugin-standard": "^2.0.1",
+    "gulp": "^3.9.1",
+    "gulp-react": "^2.0.0",
+    "vinyl-source-stream": "^1.1.0"
+  },
+  "dependencies": {
+    "react": "^15.4.2",
+    "react-dom": "^15.4.2",
+    "react-router": "^3.0.2"
+  },
+  "main": "gulpfile.js",
+  "directories": {
+    "test": "test"
+  },
+  "scripts": {
+    "test": "echo \"Error: no test specified\" && exit 1"
   }
 }
